+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog][keepachangelog] and this project adher
 
 ## Unreleased
 
+### Added
+
 - Add new state for label - text "new card"
 
 ## v1.1.0