+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this package will be documented in this file.
 
 The format is based on [Keep a Changelog][keepachangelog] and this project adheres to [Semantic Versioning][semver].
 
+## v1.1.3
+
+### Changed
+
+- Upgrade dependencies
+
 ## v1.1.2
 
 ### Changed